Fix admin statistics charts not updated on theme change
