+++ app/client/components/admin/statistics/AdminCostsBarChart.tsx
@@ -15,6 +15,7 @@ export const AdminCostsBarChart = ({
 }: AdminCostsBarChartProps) => {
   return (
     <Bar
+      redraw={true}
       data={{
         labels: costValues,
         datasets: [
